Refactoring: Add Sprockets 3 support and clean up directory structure.
This is a giant commit which brings the directory structure to be more
in line with the style of the sass-rails gem, cleans up some code
structure, and introduces support for Sprockets 3 in a way that's
cleanly separated from support for Sprockets 2.  Specifically:

  - Directory Structure: Restructure so that the main require is
    'jade/rails' instead of 'jade' as before.

  - Refactoring: Take the actual compilation code (which calls ExecJS)
    out of the root Jade module and put it into a separate Compiler
    class. This code is then invoked for both Sprockets 2 (via the Tilt
    template) and for Sprockets 3 (via the transformer).

  - Sprockets 3 Support: Add in a transformer which is structured so
    that we should actually be compatible with all Sprockets versions.
    To keep things as unchanged as possible for Sprockets 2, we'll
    continue using the old Tilt template approach, though.

NOTE: As of this commit, the ability to configure the compilation via
config.jade in Rails configuration files (e.g. config/application.rb)
has been temporarily removed. It will be added back momentarily.

For details on the Sprockets 3 compatibility/refactoring, see:
https://github.com/rails/sprockets/blob/master/guides/extending_sprockets.md